pyproject.toml
[tool.poetry]
name = "productsubscription-oc-jk"
version = "0.1.0"
description = "Yotta project 1 - product subscription"
authors = ["Jérémie Koster <jeremie.koster@gmail.com>",
    "Olivier Collier <oliviercollier8@gmail.com>"]
repository = "https://gitlab.com/yotta-academy/mle-bootcamp/projects/ml-project/winter-2021/productsubscription-oc-jk"
packages = [
    {include = "src"}
]

[tool.poetry.dependencies]
python = "^3.8"
pandas-profiling = "^2.10.1"
ipython = "^7.20.0"
scipy = "^1.6.1"
ipdb = "^0.13.4"
scikit-learn = "^0.24.1"
PyYAML = "^5.4.1"
imbalanced-learn = "^0.8.0"
category-encoders = "^2.2.2"
jupyter-contrib-nbextensions = "^0.5.1"
jupyterlab = "^3.0.10"
eli5 = "^0.11.0"
treeinterpreter = "^0.2.3"
plotly = "^4.14.3"
pytest = "^6.2.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
train = "src.application.train:main"
predict = "src.application.predict:main"
